fix: add missing font & move images
