+++ src/app/(content)/about/_assets/Ellipse 20.png
