+++ src/app/(content)/about/_assets/Rectangle 24.png
