+++ src/app/(content)/about/_assets/Rectangle 6.png
