+++ src/app/(content)/about/_assets/Rectangle 8.png
